Added api for adding comments and modifying post apis.

+++ backend/graphql_schema/schema.js
@@ -250,7 +250,7 @@ const RootQuery = new GraphQLObjectType({
           }
 
           return Users.find({username: {$regex: args.searchContent, $options: 'i'}})
-            .exec()
+            .limit(10)
             .then((data) => {
               return data;
             })
@@ -276,7 +276,7 @@ const RootQuery = new GraphQLObjectType({
     //   },
     //   async resolve(parent, args, {authUser})
     //   {
-    //     if(!authUser)
+    //     if(authUser.username !== args.username)
     //     {
     //       throw new Error("Unauthenticated user");
     //     }
@@ -548,36 +548,45 @@ const Mutation = new GraphQLObjectType({
 
         try 
         {
-          return Post.findById({_id: args.postId})
-            .exec()
-            .then((post) => {
-              if(!post) return new Error("Post doesn't exist");
+          return Users.findOne({username: args.username})
+            .then((user) => {
+              if(!user) return new Error("User does not exist");
 
-              if(args.action === "like")
-              {
-                if(post.likes.map(a=>a.liker).includes(args.username))
-                {
-                  return new Error("User " + args.username + " has already liked this post");
-                }
-                post.likes.push({liker: args.username});
-                post.save();
-                return {likeCount: post.likes.length, dislikeCount: post.dislikes.length};
-              }
-              else if(args.action === "dislike")
-              {
-                if(post.dislikes.map(a=>a.disliker).includes(args.username))
-                {
-                  return new Error("User " + args.username + " has already disliked this post");
-                }
-                post.dislikes.push({disliker: args.username});
-                post.save();
-                return {likeCount: post.likes.length, dislikeCount: post.dislikes.length};
-              }
+              return Post.findById({_id: args.postId})
+                .exec()
+                .then((post) => {
+                  if(!post) return new Error("Post doesn't exist");
+
+                  if(args.action === "like")
+                  {
+                    if(post.likes.map(a=>a.liker).includes(args.username))
+                    {
+                      return new Error("User " + args.username + " has already liked this post");
+                    }
+                    post.likes.push({liker: args.username});
+                    post.save();
+                    return {likeCount: post.likes.length, dislikeCount: post.dislikes.length};
+                  }
+                  else if(args.action === "dislike")
+                  {
+                    if(post.dislikes.map(a=>a.disliker).includes(args.username))
+                    {
+                      return new Error("User " + args.username + " has already disliked this post");
+                    }
+                    post.dislikes.push({disliker: args.username});
+                    post.save();
+                    return {likeCount: post.likes.length, dislikeCount: post.dislikes.length};
+                  }
+                })
+                .catch((err) => {
+                  console.log(err);
+                  throw err;
+                });
             })
             .catch((err) => {
               console.log(err);
               throw err;
-            })
+            });
         } 
         catch (error) 
         {
@@ -603,40 +612,49 @@ const Mutation = new GraphQLObjectType({
 
         try 
         {
-          return Post.findById({_id: args.postId})
-            .then((post) => {
-              if (!post) return new Error("Post doesn't exist");
+          return Users.findOne({username: args.username})
+            .then((user) => {
+              if(!user) return new Error("User does not exist");
 
-              if(post.posterUsername !== args.username)
-              {
-                return new Error("You are unauthorized to delete this post");
-              }
-              let imageUrl = post.image;
-              let publicId;
-              if(imageUrl !== "")
-              {
-                let splitImage = imageUrl.split('/');
-                let imageFile = splitImage[splitImage.length - 1];
-                publicId = imageFile.slice(0, imageFile.indexOf('.'));
-              }
-              
-              
-              return Post.deleteOne({_id: args.postId})
-                .then(async () => {
-                  try 
+              return Post.findById({_id: args.postId})
+                .then((post) => {
+                  if (!post) return new Error("Post doesn't exist");
+
+                  if(post.posterUsername !== args.username)
                   {
-                    if(publicId !== null && publicId !== undefined && imageUrl !== "")
-                    {
-                      await cloudinary.uploader.destroy(publicId, function(result) {console.log(result)});
-                    }
-                    return "Post deleted successfully";
-                  } 
-                  catch (error) 
+                    return new Error("You are unauthorized to delete this post");
+                  }
+                  let imageUrl = post.image;
+                  let publicId;
+                  if(imageUrl !== "")
                   {
-                    console.log(error);
-                    throw error;
+                    let splitImage = imageUrl.split('/');
+                    let imageFile = splitImage[splitImage.length - 1];
+                    publicId = imageFile.slice(0, imageFile.indexOf('.'));
                   }
                   
+                  
+                  return Post.deleteOne({_id: args.postId})
+                    .then(async () => {
+                      try 
+                      {
+                        if(publicId !== null && publicId !== undefined && imageUrl !== "")
+                        {
+                          await cloudinary.uploader.destroy(publicId, function(result) {console.log(result)});
+                        }
+                        return "Post deleted successfully";
+                      } 
+                      catch (error) 
+                      {
+                        console.log(error);
+                        throw error;
+                      }
+                      
+                    })
+                    .catch((err) => {
+                      console.log(err);
+                      throw err;
+                    });
                 })
                 .catch((err) => {
                   console.log(err);
@@ -647,6 +665,64 @@ const Mutation = new GraphQLObjectType({
               console.log(err);
               throw err;
             });
+        } 
+        catch (error) 
+        {
+          console.log(error);
+          throw error;
+        }
+      }
+    },
+
+    // All APIs related to comments are below
+
+    // Add comments to a particular post
+    addComments: {
+      type: CommentsInputType,
+      args: {
+        username: { type: new GraphQLNonNull(GraphQLString) },
+        postId: { type: new GraphQLNonNull(GraphQLID) },
+        commentContent: { type: new GraphQLNonNull(GraphQLString) }
+      },
+      async resolve(parent, args, {authUser})
+      {
+        if(authUser.username !== args.username)
+        {
+          return new Error("Unauthenticated user");
+        }
+
+        if(args.commentContent === "")
+        {
+          return new Error("Comment cannot be empty");
+        }
+
+        try 
+        {
+          return Users.findOne({username: args.username})
+          .then((user) => {
+            if(!user) return new Error("User does not exist");
+
+            return Post.findById({_id: args.postId})
+              .then((post) => {
+                if (!post) return new Error("Post doesn't exist");
+
+                let newComment = {commenter: args.username, commentContent: args.commentContent};
+                post.comments.push(newComment);
+                post.save();
+                let date = post.comments[post.comments.length - 1].commentDate;
+                let returnComment = {commenter: args.username, commentContent: args.commentContent, 
+                  commentDate: date};
+                return returnComment;
+              })
+              .catch((err) => {
+                console.log(err);
+                throw err;
+              });
+          })
+          .catch((err) => {
+            console.log(err);
+            throw err;
+          });
         } 
         catch (error) 
         {
